ml/requirements.txt
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
